+++ NEWS.md
@@ -2,3 +2,4 @@
 
 * Added a `NEWS.md` file to track changes to the package.
 * Added `plot_map()` function to create Norwegian choropleth municipality and country maps 
+* Added extra check in `synthetic_data()`, due to some incompatibility issues with KlassR it currently requires stringr v1.5.1